refactor(stella-protocol): split event.rs's payload types so the wire contract clears the ratchet (#3436)
`crates/stella-protocol/src/event.rs` had reached 1533 lines — over the
1500-line limit and deliberately not grandfathered, so `make file-size` was
red against `main` for every branch whose merge base predated the growth.

The seam is the one the file already uses twice. Its tests were split to a
sibling for exactly this reason (#1857), and `crate::receipt`'s types are
re-exported from here so consumers keep resolving. The leaf payload types a
variant *carries* — `FileChangeKind`, `VerdictEvidence`, `ScopeProposal`,
`HunkProposal`, `ProposedHunk`, `MediaKind`, `MediaJobState`,
`MediaArtifactRef`, `PrStatus`, `CiStatus`, `TaskItem`, `TaskStatus` — move
to `event/payload.rs` and are re-exported unchanged.

Nothing moved is a variant, so the totality machinery stays untouched:
`event/tags.rs` and `event/consumers.rs` still generate `KNOWN_TYPE_TAGS`,
`type_tag()` and the consumer ledger from one table, and invariant #10
remains a compile-time `E0004`. `AgentEvent` itself, its `Serialize`/
`Deserialize` impls, and the stage/budget/policy enums all stay put.

No behavior change and no wire change: `make wire-schema` reports
`docs/wire/` still matches the types, and every path is a re-export, so
`stella_protocol::event::TaskItem` resolves exactly as before.

event.rs: 1533 -> 1249. No baseline entry was added — this file predates
nothing and the baseline is for files that predate the guard.

Refs #3436
